Add files via upload
I've improved a full-view mosaic program on the github. In the original result, there will be a light-and-dark difference in the results of the two images. I added two functions to the back of the panorama.py file, which can be a better image fusion to remove the dark and dark.